regional/requirements.txt
# 核心依赖
loguru==0.7.2
paho-mqtt==2.1.0
pika==1.3.2
python-dotenv==1.0.0

# 可选依赖（如果需要 HTTP 接口）
# Flask==3.0.3
# Flask-Cors==4.0.0
# flasgger==0.9.7.1

# 工具依赖
requests==2.32.2
PyYAML==6.0.2




